Checked is working, added todos. Split library into source and header.
